Rearrange the files
